requirements.txt
# Core web framework
fastapi>=0.116.1
uvicorn>=0.35.0
streamlit>=1.47.1

# Embeddings and LLM tooling
sentence-transformers>=5.0.0
langchain>=0.3.26
openai>=1.97.1
google-generativeai>=0.8.5

# Document handling
python-docx>=1.2.0
pymupdf>=1.26.3
python-multipart>=0.0.20
unstructured>=0.18.9

# Utilities
faiss-cpu>=1.11.0.post1
numpy>=2.3.1
pandas>=2.3.1
pydantic>=2.11.7
tiktoken>=0.9.0
